+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Old events are now filtered when listing
 - Display a hint when nothing is found during listing.
 - Enable colored output for clap
+- Add default priority configuration
 
 ### Fixed